split up test-in-memory into seperate files
