+++ packages/integration-tests/src/test-in-memory/data_storage_auth_with_triggers.test.ts
@@ -1,11 +1,9 @@
-import { dataStorageAuthWithTriggers } from '../test-projects/data-storage-auth-with-triggers-ts/amplify/test_factories.js';
+import { it } from 'node:test';
 import {
   assertExpectedLogicalIds,
   synthesizeBackendTemplates,
 } from '../define_backend_template_harness.js';
-import { it } from 'node:test';
-import { dataWithoutAuth } from '../test-projects/standalone-data-auth-modes/amplify/test_factories.js';
-import { dataWithoutAuthNoAuthMode } from '../test-projects/standalone-data-sandbox-mode/amplify/test_factories.js';
+import { dataStorageAuthWithTriggers } from '../test-projects/data-storage-auth-with-triggers-ts/amplify/test_factories.js';
 
 /**
  * This test suite is meant to provide a fast feedback loop to sanity check that different feature verticals are working properly together.
@@ -59,35 +57,4 @@ void it('data storage auth with triggers', () => {
     'onDeletelambda96BB6F15',
   ]);
   /* eslint-enable spellcheck/spell-checker */
-});
-
-void it('data without auth with lambda auth mode', () => {
-  const templates = synthesizeBackendTemplates(dataWithoutAuth);
-
-  assertExpectedLogicalIds(templates.root, 'AWS::CloudFormation::Stack', [
-    'data7552DF31',
-    'function1351588B',
-  ]);
-  assertExpectedLogicalIds(templates.data, 'AWS::AppSync::GraphQLApi', [
-    'amplifyDataGraphQLAPI42A6FA33',
-  ]);
-  assertExpectedLogicalIds(templates.data, 'AWS::CloudFormation::Stack', [
-    'amplifyDataAmplifyTableManagerNestedStackAmplifyTableManagerNestedStackResource86290833',
-    'amplifyDataTodoNestedStackTodoNestedStackResource551CEA56',
-  ]);
-});
-
-void it('data without auth with default auth mode', () => {
-  const templates = synthesizeBackendTemplates(dataWithoutAuthNoAuthMode);
-
-  assertExpectedLogicalIds(templates.root, 'AWS::CloudFormation::Stack', [
-    'data7552DF31',
-  ]);
-  assertExpectedLogicalIds(templates.data, 'AWS::AppSync::GraphQLApi', [
-    'amplifyDataGraphQLAPI42A6FA33',
-  ]);
-  assertExpectedLogicalIds(templates.data, 'AWS::CloudFormation::Stack', [
-    'amplifyDataAmplifyTableManagerNestedStackAmplifyTableManagerNestedStackResource86290833',
-    'amplifyDataTodoNestedStackTodoNestedStackResource551CEA56',
-  ]);
 });
